fix: move code to proper location
